editoast: move OperationalPointCache to infra_cache
