+++ editoast/src/infra_cache/mod.rs
@@ -32,6 +32,7 @@ use thiserror::Error;
 use crate::error::Result;
 use crate::infra_cache::object_cache::BufferStopCache;
 use crate::infra_cache::object_cache::DetectorCache;
+use crate::infra_cache::object_cache::OperationalPointCache;
 use crate::infra_cache::operation::CacheOperation;
 use crate::infra_cache::operation::RailjsonObject;
 use crate::map::BoundingBox;
@@ -738,6 +739,7 @@ pub mod tests {
     use super::TrackSectionCache;
     use crate::infra_cache::object_cache::BufferStopCache;
     use crate::infra_cache::object_cache::DetectorCache;
+    use crate::infra_cache::object_cache::OperationalPointPartCache;
     use crate::infra_cache::operation::create::tests::create_buffer_stop;
     use crate::infra_cache::operation::create::tests::create_detector;
     use crate::infra_cache::operation::create::tests::create_electrification;
@@ -754,7 +756,6 @@ pub mod tests {
     use crate::modelsv2::infra::tests::test_infra_transaction;
     use crate::schema::Electrification;
     use crate::schema::OperationalPoint;
-    use crate::schema::OperationalPointPartCache;
     use crate::schema::Route;
     use crate::schema::SpeedSection;
     use crate::schema::Switch;
