+++ editoast/src/views/infra/auto_fixes/operational_point.rs
@@ -1,14 +1,14 @@
+use editoast_schemas::primitives::OSRDObject as _;
+use editoast_schemas::primitives::ObjectRef;
 use std::collections::HashMap;
 
 use tracing::debug;
 
 use super::new_ref_fix_delete_pair;
 use super::Fix;
+use crate::infra_cache::object_cache::OperationalPointCache;
 use crate::schema::InfraError;
 use crate::schema::InfraErrorType;
-use crate::schema::OperationalPointCache;
-use editoast_schemas::primitives::OSRDObject as _;
-use editoast_schemas::primitives::ObjectRef;
 
 pub fn fix_operational_point(
     operational_point: &OperationalPointCache,
